Move PlayPauseButton.vue under src/components/
