+++ src/components/PlayPauseButton.vue
